my_ls with a ~few~ bugs...
