+++ countEntries.c
@@ -1,0 +1,10 @@
+#include "my_ls.h"
+
+int countEntries(entryList* head){
+    int fileCount = 0;
+    while(head != NULL){
+        fileCount++;
+        head = head->next;
+    }
+    return fileCount;
+}
